# Cache unnormalized intersection types