@ahejlsberg 2019-05-20
@typescript-bot run dt