@ahejlsberg 2019-05-20
> Should we consider doing the same for unions ... ?

I tried. It doesn't seem to make much of a difference. In fact, since `UnionReduction.Literal` is fairly efficient the caching actually had a negative effect (because computing keys takes time). But even when just caching `UnionReduction.Subtype` there wasn't a measurable effect.